CIV-3388
Update logic of Refer handler - move it under one handler.
