+++ Jenkinsfile_CNP
@@ -9,7 +9,7 @@ def product = "civil"
 def component = "general-applications"
 def camundaBranch = "master"
 
-def generalappCCDBranch = "sprint6integration"
+def generalappCCDBranch = "feature/CIV-3388"
 
 AppPipelineConfig pipelineConf
